Moved AlgorandAccount types to types.ts
