+++ packages/core/src/lib/chains/algorand/AccountsStore.ts
@@ -6,6 +6,7 @@ import algosdk, {
 } from "algosdk";
 import SendRawTransaction from "algosdk/dist/types/client/v2/algod/sendRawTransaction";
 import {AlgorandStandardAssetConfig} from "src/lib/common";
+import {AlgorandAccountInformation} from "./types";
 
 export type AlgorandAccount = {
   addr: string;
@@ -20,31 +21,6 @@ export type AlgorandMultiSigAccount = {
   addresses: string[];
   params: MultisigMetadata;
   information: AlgorandAccountInformation;
-};
-
-export type AlgorandAccountInformation = {
-  address: string;
-  amount: number;
-  "amount-without-pending-rewards": number;
-  "apps-local-state": Array<any>;
-  "apps-total-schema": { "num-byte-slice": number; "num-uint": number };
-  "created-apps": Array<any>;
-  "created-assets": Array<any>;
-  "min-balance": number;
-  "pending-rewards": number;
-  "reward-base": number;
-  rewards: number;
-  round: number;
-  status: string;
-  "total-apps-opted-in": number;
-  "total-assets-opted-in": number;
-  "total-created-apps": number;
-  "total-created-assets": number;
-  assets: Array<{
-    amount: number;
-    "asset-id": number;
-    "is-frozen": boolean;
-  }>;
 };
 
 export class AlgorandAccountsStore {
